+++ src/Orchestra/Notifier/NotifierInterface.php
@@ -1,17 +1,16 @@
 <?php namespace Orchestra\Notifier;
 
-use Illuminate\Auth\Reminders\RemindableInterface;
 
 interface NotifierInterface
 {
     /**
      * Send notification via API.
      *
-     * @param  \Illuminate\Auth\Reminders\RemindableInterface  $user
-     * @param  string                                          $subject
-     * @param  string|array                                    $view
-     * @param  array                                           $data
+     * @param  NotifiableInterface $user
+     * @param  string              $subject
+     * @param  string|array        $view
+     * @param  array               $data
      * @return boolean
      */
-    public function send(RemindableInterface $user, $subject, $view, array $data = array());
+    public function send(NotifiableInterface $user, $subject, $view, array $data = array());
 }